Remove validators from ScopedKey to avoid too much thrashing
Also:
- moved models tests to to their own unit test module.
- added wildcard scopes to `test_base_api::tokendata
- added casting of '*' as inputs to `Scope` to `None`; this is our
  representation of wildcards for `Scope` components.
