+++ fah_alchemy/tests/unit/test_base_api.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException
 from pydantic import ValidationError
 
 from fah_alchemy.base.api import validate_scopes, validate_scopes_query
-from fah_alchemy.models import Scope, ScopedKey, InvalidScopeError
+from fah_alchemy.models import Scope, ScopedKey
 from fah_alchemy.security.models import Token, TokenData, CredentialedEntity
 
 
@@ -26,6 +26,8 @@ def tokendata():
             "org1-campaignA-projectI",
             "org1-campaignB-projectI",
             "org1-campaignB-projectII",
+            "org2-*-*",
+            "org3-campaignA-*",
         ],
     )
 
@@ -77,19 +79,3 @@ def test_validate_scopes_query(tokendata):
 
     assert len(matches) == 3
     assert matches == tokendata.scopes
-
-
-@pytest.mark.parametrize(
-    "scope_string", ["*-*-*", "org1-*-*", "org1-campaignA-*", "org1-campaignA-projectI"]
-)
-def test_wildcard_scopes_valid(scope_string):
-    scope = Scope.from_str(scope_string)
-
-
-@pytest.mark.parametrize(
-    "scope_string",
-    ["*-*-projectI", "*-campaignA-*", "*-campaignA-projectI", "org1-*-projectI"],
-)
-def test_wildcard_scopes_invalid(scope_string):
-    with pytest.raises(ValidationError):
-        scope = Scope.from_str(scope_string)
